requirements.txt
h5py==2.10.0
numpy
tensorflow-gpu==1.14.0
opencv-python==4.1.1.26
keras==2.2.5
Pillow
imageio
sklearn
scikit-learn
matplotlib
cryptography
segmentation_models
pytest-shutil
community